# frontend/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Crystalith",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/notebooks/{notebook_id}/refine": {
      "post": {
        "tags": [
          "refine"
        ],
        "summary": "Refine",
        "operationId": "refine_v1_notebooks__notebook_id__refine_post",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notebook Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/notebooks/{notebook_id}/refine/batch": {
      "post": {
        "tags": [
          "refine"
        ],
        "summary": "Refine Batch",
        "operationId": "refine_batch_v1_notebooks__notebook_id__refine_batch_post",
        "parameters": [
          {
            "name": "notebook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notebook Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefineBatchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefineBatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Citation": {
        "properties": {
          "source_id": {
            "type": "integer",
            "title": "Source Id"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "chunk_id": {
            "type": "integer",
            "title": "Chunk Id"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "paragraph_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph Index"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "source_name",
          "chunk_id",
          "chunk_index",
          "snippet"
        ],
        "title": "Citation"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RefineBatchOutput": {
        "properties": {
          "paragraph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph"
          },
          "bullets": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bullets"
          },
          "structured": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructuredRefine"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "RefineBatchOutput"
      },
      "RefineBatchRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          },
          "formats": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Formats"
          },
          "chunk_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Ids"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "min_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Score",
            "default": 0.2
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "RefineBatchRequest"
      },
      "RefineBatchResponse": {
        "properties": {
          "outputs": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RefineBatchOutput"
            },
            "type": "object",
            "title": "Outputs"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "evidence": {
            "type": "boolean",
            "title": "Evidence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "outputs",
          "citations",
          "evidence",
          "created_at"
        ],
        "title": "RefineBatchResponse"
      },
      "RefineRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "minLength": 1,
            "title": "Prompt"
          },
          "format": {
            "type": "string",
            "title": "Format",
            "default": "paragraph"
          },
          "chunk_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Ids"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "default": 5
          },
          "min_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Score",
            "default": 0.2
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "RefineRequest"
      },
      "RefineResponse": {
        "properties": {
          "format": {
            "type": "string",
            "title": "Format"
          },
          "paragraph": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paragraph"
          },
          "bullets": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bullets"
          },
          "structured": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StructuredRefine"
              },
              {
                "type": "null"
              }
            ]
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          },
          "evidence": {
            "type": "boolean",
            "title": "Evidence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "format",
          "citations",
          "evidence",
          "created_at"
        ],
        "title": "RefineResponse"
      },
...
      "StructuredRefine": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "bullets": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bullets"
          },
          "terms": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Terms"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/Citation"
            },
            "type": "array",
            "title": "Citations"
          }
        },
        "type": "object",
        "required": [
          "title",
          "bullets",
          "terms",
          "citations"
        ],
        "title": "StructuredRefine"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}